Move lpGkDim from dyn module to kernel/combinatorics sucht that it can be invoked with the native `dim` command
