+++ kernel/combinatorics/stairc.h
@@ -31,6 +31,11 @@ void scDegree(ideal  s,intvec *modulweight,ideal Q=NULL);
 
 ideal scKBase(int deg, ideal  s, ideal Q=NULL, intvec * mv=NULL);
 
+#if HAVE_SHIFTBBA
+int lp_gkDim(const ideal G);
+intvec * lp_ufnarovskiGraph(ideal G, ideal &standardWords);
+#endif
+
 #endif
